requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
yarl==1.5.1aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
iniconfig==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
multidict==4.7.6
packaging==20.4
pep8==1.7.1
pluggy==0.13.1
py==1.9.0
pylint==2.5.3
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
rope==0.17.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
yarl==1.5.1